pyproject.toml
[tool.poetry]
name = "teste-hl"
version = "0.1.0"
description = ""
authors = ["MatheusFariasSousa <143126173+MatheusFariasSousa@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
pytest = "^8.2.2"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
pyjwt = "^2.8.0"
pwdlib = {extras = ["argon2"], version = "^0.2.0"}
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
